Starting to refactor the book a bit.
Splitting up some examples, using new names, properly testing the book build (by ignoring unused files), adding more explicit
introductions and authors etc.
