+++ .github/workflows/book_stable.yml
@@ -29,7 +29,7 @@ jobs:
       - uses: actions/checkout@v4
 
       - name: Build the book
-        run: jupyter-book build .
+        run: jupyter-book build . -W --keep-going
 
       - uses: actions/upload-artifact@v4
         if : always()
